Minor update v6.3.0

+++ src/disturbances/air_drag.cpp
@@ -17,7 +17,7 @@ AirDrag::AirDrag(const std::vector<Surface>& surfaces, const libra::Vector<3>& c
       wall_temperature_K_(wall_temperature_K),
       molecular_temperature_K_(molecular_temperature_K),
       molecular_weight_g_mol_(molecular_weight_g_mol) {
-  int num = surfaces_.size();
+  size_t num = surfaces_.size();
   ct_.assign(num, 1.0);
   cn_.assign(num, 0.0);
 }
