+++ src/library/initialize/initialize_file_access.hpp
@@ -138,7 +138,7 @@ class IniAccess {
    * @param[out] output_value: Read double matrix value
    * @param[in] node_num: Number of node. When reading n * m matrix, please substitute bigger number.
    */
-  void ReadCsvDouble(std::vector<std::vector<double>>& output_value, const int node_num);
+  void ReadCsvDouble(std::vector<std::vector<double>>& output_value, const size_t node_num);
   /**
    * @fn ReadCsvDoubleWithHeader
    * @brief Read matrix value in CSV file with header
@@ -155,7 +155,7 @@ class IniAccess {
    * @param[out] output_value: Read matrix of string
    * @param[in] node_num: Number of node. When reading n * m matrix, please substitute bigger number.
    */
-  void ReadCsvString(std::vector<std::vector<std::string>>& output_value, const int node_num);
+  void ReadCsvString(std::vector<std::vector<std::string>>& output_value, const size_t node_num);
 
  private:
   static const size_t kMaxCharLength = 1024;
